thaw UI & toasts
