+++ crates/rust-sample-wallet/Cargo.toml
@@ -8,7 +8,7 @@ console_error_panic_hook = "0.1.7"
 tracing = "0.1.41"
 tracing-subscriber = "0.3.19"
 tracing-subscriber-wasm = "0.1.0"
-leptos = { version = "0.8.2", features = ["csr"] }
+leptos = { version = "0.8.5", features = ["csr"] }
 leptos-use = "0.16.2"
 serde.workspace = true
 serde_json.workspace = true
@@ -18,3 +18,4 @@ yttrium = { path = "../yttrium", default-features = false, features = [
     "sign_client",
 ] }
 web-sys = { version = "0.3.77", features = ["Window", "Storage"] }
+thaw = { git = "https://github.com/thaw-ui/thaw", branch = "thaw/v0.5", features = ["csr"] }
